requirements.txt
python-telegram-bot[socks]

